examples/md_to_pdf_agent/pyproject.toml
[project]
name = "md-to-pdf"
version = "0.1.0"
description = "Markdown to PDF conversion agent example"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-anthropic>=0.3.6",
    "langchain-core>=0.3.31",
    "middleman-ai",
    "python-dotenv>=1.0.1",
]

[tool.uv]
[tool.uv.sources]
middleman-ai = { path = "../../", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "src/md_to_pdf",
]
